+++ C#/AutoHotInterception/Helpers/HelperFunctions.cs
@@ -46,7 +46,7 @@ namespace AutoHotInterception.Helpers
         /// <param name="btn">0 = LMB, 1 = RMB etc</param>
         /// <param name="state">1 = Press, 0 = Release</param>
         /// <returns>A State value for a Mouse Stroke</returns>
-        public static ManagedWrapper.Stroke ButtonAndStateToStroke(int btn, int state)
+        public static ManagedWrapper.Stroke MouseButtonAndStateToStroke(int btn, int state)
         {
             var stroke = new ManagedWrapper.Stroke();
             var power = btn < 5 ? btn * 2 + (state == 0 ? 1 : 0) : btn + 5;
@@ -55,7 +55,7 @@ namespace AutoHotInterception.Helpers
             return stroke;
         }
 
-        public static ButtonState StrokeStateToButtonState(ManagedWrapper.Stroke stroke)
+        public static ButtonState MouseStrokeToButtonState(ManagedWrapper.Stroke stroke)
         {
             int state = stroke.mouse.state;
             ushort btn = 0;
@@ -79,6 +79,50 @@ namespace AutoHotInterception.Helpers
             return new ButtonState {Button = btn, State = state};
         }
 
+        public static KeyboardState KeyboardStrokeToKeyboardState(ManagedWrapper.Stroke stroke)
+        {
+            var code = stroke.key.code;
+            var state = stroke.key.state;
+            var retVal = new KeyboardState();
+            if (code == 54) code = 310;
+
+            // If state is shifted up by 2 (1 or 2 instead of 0 or 1), then this is an "Extended" key code
+            if (state > 1)
+            {
+                if (code == 42 || state > 3)
+                {
+                    // code == 42
+                    // Shift (42/0x2a) with extended flag = the key after this one is extended.
+                    // Example case is Delete (The one above the arrow keys, not on numpad)...
+                    // ... this generates a stroke of 0x2a (Shift) with *extended flag set* (Normal shift does not do this)...
+                    // ... followed by 0x53 with extended flag set.
+                    // We do not want to fire subsriptions for the extended shift, but *do* want to let the key flow through...
+                    // ... so that is handled here.
+                    // When the extended key (Delete in the above example) subsequently comes through...
+                    // ... it will have code 0x53, which we shift to 0x153 (Adding 256 Dec) to signify extended version...
+                    // ... as this is how AHK behaves with GetKeySC()
+
+                    // state > 3
+                    // Pause sends code 69 normally as state 0/1, but also LCtrl (29) as state 4/5
+                    // Ignore the LCtrl in this case
+
+                    // Set flag to indicate ignore
+                    retVal.Ignore = true;
+                }
+                else
+                {
+                    // Extended flag set
+                    // Shift code up by 256 (0x100) to signify extended code
+                    code += 256;
+                    state -= 2;
+                }
+            }
+
+            retVal.Code = code;
+            retVal.State = (ushort) (1 - state);
+            return retVal;
+        }
+
         public class DeviceInfo
         {
             public int Id { get; set; }
@@ -92,6 +136,13 @@ namespace AutoHotInterception.Helpers
         {
             public ushort Button { get; set; }
             public int State { get; set; }
+        }
+
+        public class KeyboardState
+        {
+            public ushort Code { get; set; }
+            public ushort State { get; set; }
+            public bool Ignore { get; set; }
         }
     }
 }